@@ -5,18 +5,23 @@
   "contributors": [
     {
       "name": "Tommy Gaessler"
+    },
+    {
+      "name": "Saulo Costa",
+      "url": "https://github.com/saulotarsobc"
     }
   ],
   "homepage": "",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@zoom/meetingsdk": "^3.6.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "@zoom/meetingsdk": "^3.6.1",
+    "jsrsasign": "^11.1.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
